created generalFiles package (+ refactor)
