+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/community/post/generalFiles/GeneralFilesListItem.kt
@@ -1,4 +1,4 @@
-package com.sju18001.petmanagement.ui.community.post
+package com.sju18001.petmanagement.ui.community.post.generalFiles
 
 data class GeneralFilesListItem (
     val postId: Long,
