+++ client/android/app/src/main/res/layout/activity_general_files.xml
@@ -4,7 +4,7 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".ui.community.post.GeneralFilesActivity">
+    tools:context=".ui.community.post.generalFiles.GeneralFilesActivity">
 
     <RelativeLayout
         android:layout_width="wrap_content"
